Move to data-raw/ the main table used to gather data.
This table is moved out of data/ because it is not exported. Instead, it is used to generate exported data that are subsets of this table.

To learn details about exported data see http://r-pkgs.had.co.nz/data.html#data-data.

To learn details about data-raw/ see http://r-pkgs.had.co.nz/data.html and search for "data-raw".