Use continuous integration with GitHub actions
This commit adds CI with GitHub Actions in preparation for a new
version of rOpenSci's dev guide that asks contributors to move
away from Travis CI.

It also add badge to README and render it.

The developers guide of rOpenSci states that a continous integration
service is required. As of today, the guide suggests Travis but
rOpenSci is about to remove mentions of Travis CI in favor of
GitHub actions.

See https://devdevguide.netlify.app/booknews.html; it says:
2021-01-12, remove mentions of Travis CI and link to Jeroen Ooms’ blog
post about moving away from Travis.