Increment version number
The comment following review asks to increment the version number,
and to list all changes made.

I did increment the version number (with `usethis::use_version()`)
and summarized all changes with a reference to the peer-review
issue. I guess we could do better than this, but this seems better
than nothing for now.

Would it be possible to add some detail, maybe in the form of
bullets under my general comment? Maybe that's a good place to
@mention the reviewers and authors, as a way to acknowledge their
help.

Note we can offer reviewers (not editors) to add them as reviewers
in the DESCRIPTION file. I don't thik they would show up in the
citation but it's still a way for them to prove their contribution
to research software.