Avoid spelling note in R CMD check
The word Extratropical gets picked up by the automated spellchecker
during R CMD check. This rises a note the will likely result in
R core team asking us to fix the spelling. Even if we argue for
leaving the word as it, it may take an extra round of review. To
increase the odds the package will be accepted as is I suggest
using extra-tropical instead. This applies only to DESCRIPTION and
likely we can change it back to extratropical after the fist release
and with no consequences.