Test errors of `new_equations()` and flag FIXME
Here I test that `new_equations()` fails gracefully. Along the
way I noted unexpected behaviour that I captured in the form of
tests in tests/testthat/test-FIXME.R

TODO:

* Continue testing each call to stop() behaves as expected. See
  how I did it in tests/testthat/test-new_equations.R. Note I
  call expect_snapshot_error() so we can see the exact error
  message the user reads (see tests/testthat/_snaps). I welcome
  help as this function is quite tedious to test (see #215).

A good way to note which lines need tests is with a report of
test coverage. In RStudio go to the Addins menu and type "report".
You should see an adding from the devtools packages named
"Report test coverage for a package". Search for lines highlighted
in red and test them so they turn green. Once you're done the
package will be 100% covered :=)